requirements.txt
fastapi
uvicorn[standard]
pytest
httpx
pydantic-settings
SQLAlchemy
python-dotenv
structlog
scapy
